requirements.txt
fastapi
farm-haystack[preprocessing,elasticsearch,inference,metrics]
uvicorn
torch
torchvision
torchaudio
streamlit
Markdown
st-annotated-text
python-multipart
pyperclip